Video Apis: added flag reflective quiz status = completed or incomplete
